Pre comp1

+++ NewBasicTemplate2016/src/Commands/AutomationOfArm.h
@@ -7,7 +7,7 @@
 class AutomationOfArm: public CommandBase
 {
 private:
-	double currentValue,
+	int currentValue,
 		   desiredValue,
 		   startingValue,
 		   angle;
